Major Update:   See details
Roles have been added to the AllUsers Table to help with quickly filtering Users by roles (faster queries and cleaner code).

Clerk functions filter Users by Role in order to Assign on Students in Clerk-related Departments to Subjects in Clerk-Related Departments (Enrollments)

Same for associating Examners to Subjects  (Theacings)

When an Admin adds a new User, the Role is added in the AllUsers Table.

Student Dashboard links to functions to View the Time Table and Results for Exam.
The Student Can apply for Resit if Assessment==Failed.  This WILL RESET THE ENROLLMENT.

Some changes to the Automatic correction and production of stats had to made in order for the above to work.

These latter changes solve a big bug.  In short now we can have resits without the Student having marks from previous exams being counted with the resit.

Admin and Clerk can add exam Locations to the Database.

Many actions have been locked so that they can only be triggered by users in the authenticated with the right role