DateTime Picker, Full implementation of Access Code and Logs
 Managed to add the  Date and Time Picker for the Clerk to set DateTime for the Exam.   It is not friendly because of the issue mentioned above;  I found a way to minimise the problem, it's usable but not as tidy as one would wish for.

This picker will vary depending on the browser used, because I am not using JQuery UI.

I also added started security around the exam.  The Student now has to input the Access Code within 6 minutes of issue.  All attempts will be logged in the DB (new table new migration).

As far as functionality, all that remains is related to photos.  ( I still do not know how the "ID Photo" of the student will be added)

I will not update to Git for now because I want to try to improve the DateTime thing and test other stuff more thoroughly.